Move memory serialization in the corresponding header away from types.h
