+++ ttnn/cpp/ttnn/tensor/serialization.hpp
@@ -9,9 +9,7 @@
 #include <string>
 #include <unordered_map>
 
-namespace tt {
-
-namespace tt_metal {
+namespace tt::tt_metal {
 
 void dump_tensor(
     const std::string& file_name, const Tensor& tensor, const std::unordered_map<std::string, std::string>& strategy);
@@ -19,6 +17,10 @@ void dump_tensor(
 Tensor load_tensor(const std::string& file_name, Device* device = nullptr);
 Tensor load_tensor(const std::string& file_name, distributed::MeshDevice* device = nullptr);
 
-}  // namespace tt_metal
+void dump_memory_config(std::ostream& output_stream, const MemoryConfig& memory_config);
+void dump_memory_config(const std::string& file_name, const MemoryConfig& memory_config);
 
-}  // namespace tt
+MemoryConfig load_memory_config(std::ifstream& input_stream);
+MemoryConfig load_memory_config(const std::string& file_name);
+
+}  // namespace tt::tt_metal
