+++ ttnn/cpp/ttnn/tensor/types.hpp
@@ -280,12 +280,6 @@ std::ostream& operator<<(std::ostream& os, const MemoryConfig& config);
 bool operator==(const MemoryConfig &config_a, const MemoryConfig &config_b);
 bool operator!=(const MemoryConfig &config_a, const MemoryConfig &config_b);
 
-void dump_memory_config(std::ostream &output_stream, const MemoryConfig &memory_config);
-void dump_memory_config(const std::string &file_name, const MemoryConfig &memory_config);
-
-MemoryConfig load_memory_config(std::ifstream &input_stream);
-MemoryConfig load_memory_config(const std::string &file_name);
-
 using OwnedBuffer = std::variant<
     owned_buffer::Buffer<uint8_t>,
     owned_buffer::Buffer<uint16_t>,
